refactor: 부동소수점 Double 타입으로 변환

+++ src/main/java/com/econo_4factorial/newproject/base/domain/Base.java
@@ -30,7 +30,7 @@ public class Base {
 
     private String weather;
 
-    private Float temperature;
+    private Double temperature;
 
     @Column(precision = 16, scale = 14, nullable = false)
     private BigDecimal latitude;
